# apps/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "RevelAI API",
    "version": "0.0.0"
  },
  "paths": {
...
    "/v1/challenges": {
      "get": {
        "operationId": "listChallenges",
        "summary": "List active challenges",
        "parameters": [],
        "responses": {
          "200": {
            "description": "HTTP 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeListResponse"
                }
              }
            }
          },
          "400": {
            "description": "HTTP 400",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteError"
                }
              }
            }
          },
          "503": {
            "description": "HTTP 503",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/calibration-sessions": {
      "post": {
        "operationId": "createCalibrationSession",
        "summary": "Issue a calibration session",
        "security": [
          {
            "AthleteIdentity": []
          }
        ],
        "parameters": [
          {
            "name": "X-RevelAI-Athlete-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalibrationSessionCreateInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "HTTP 201",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalibrationSession"
                }
              }
            }
          },
          "400": {
            "description": "HTTP 400",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteError"
                }
              }
            }
          },
          "503": {
            "description": "HTTP 503",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/attempts": {
      "get": {
        "operationId": "listAttempts",
        "summary": "List attempts for the local athlete",
        "security": [
          {
            "AthleteIdentity": []
          }
        ],
        "parameters": [
          {
            "name": "X-RevelAI-Athlete-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "default": 20,
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "HTTP 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptListResponse"
                }
              }
            }
          },
          "400": {
            "description": "HTTP 400",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteError"
                }
              }
            }
          },
          "503": {
            "description": "HTTP 503",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteError"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createAttempt",
        "summary": "Create a Free or Verified attempt",
        "security": [
          {
            "AthleteIdentity": []
          }
        ],
        "parameters": [
          {
            "name": "X-RevelAI-Athlete-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAttemptInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "HTTP 201",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAttemptResponse"
                }
              }
            }
          },
          "400": {
            "description": "HTTP 400",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteError"
                }
              }
            }
          },
          "404": {
            "description": "HTTP 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteError"
                }
              }
            }
          },
          "409": {
            "description": "HTTP 409",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteError"
                }
              }
            }
          },
          "410": {
            "description": "HTTP 410",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteError"
                }
              }
            }
          },
          "503": {
            "description": "HTTP 503",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/attempts/{id}/result": {
      "get": {
        "operationId": "getAttemptResult",
        "summary": "Read an attempt outcome",
        "security": [
          {
            "AthleteIdentity": []
          }
        ],
        "parameters": [
          {
            "name": "X-RevelAI-Athlete-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "HTTP 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptResultResponse"
                }
              }
            }
          },
          "202": {
            "description": "HTTP 202",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptResultResponse"
                }
              }
            }
          },
          "400": {
            "description": "HTTP 400",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteError"
                }
              }
            }
          },
          "404": {
            "description": "HTTP 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteError"
                }
              }
            }
          },
          "503": {
            "description": "HTTP 503",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AthleteIdentity": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RevelAI-Athlete-Id"
      }
    },
    "schemas": {
...
      "CalibrationSessionCreateInput": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "challengeId": {
            "type": "string",
            "const": "wall-pass"
          },
          "challengeVersion": {
            "type": "number",
            "const": 1
          }
        },
        "required": [
          "challengeId",
          "challengeVersion"
        ],
        "additionalProperties": false
      },
...
      "ChallengeListResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "const": "wall-pass"
                },
                "version": {
                  "type": "number",
                  "const": 1
                },
                "sport": {
                  "type": "string",
                  "const": "futsal"
                },
                "activeDurationSeconds": {
                  "type": "number",
                  "const": 60
                },
                "calibrationPreRollSeconds": {
                  "type": "number",
                  "const": 4
                },
                "requiredGates": {
                  "type": "array",
                  "prefixItems": [
                    {
                      "type": "string",
                      "const": "device"
                    },
                    {
                      "type": "string",
                      "const": "space"
                    },
                    {
                      "type": "string",
                      "const": "athlete"
                    },
                    {
                      "type": "string",
                      "const": "rehearsal"
                    },
                    {
                      "type": "string",
                      "const": "record"
                    }
                  ]
                }
              },
              "required": [
                "id",
                "version",
                "sport",
                "activeDurationSeconds",
                "calibrationPreRollSeconds",
                "requiredGates"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "items"
        ],
        "additionalProperties": false
      },
...
    }
  }
}